refactor(github): split issue read plumbing
